% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{CHTpackage}
\name{CHTpackage}
\alias{CHTpackage}
\title{CHTpackage: Algorithms for Stable Matching and Blood Type Compatibility}
\description{
CHTpackage includes implementations of algorithms for stable matching and blood type compatibility.
}
