% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_bucket_complexity_test_function_cpp.R
\name{test_gs_bucket_time_cpp}
\alias{test_gs_bucket_time_cpp}
\title{Test the runtime of Gale-Shapley with bucket cpp version}
\usage{
test_gs_bucket_time_cpp(n)
}
\arguments{
\item{n}{Number of men/women.}
}
\value{
Execution time in seconds.
}
\description{
Runs the algorithm for random preference lists of size n and returns execution time.
}
