% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_complexity_test_function_r.R
\name{test_gs_time}
\alias{test_gs_time}
\title{Test the runtime of Gale-Shapley}
\usage{
test_gs_time(n)
}
\arguments{
\item{n}{Number of men/women.}
}
\value{
Execution time in seconds.
}
\description{
Runs the algorithm for random preference lists of size n and returns execution time.
}
